% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{multibridge_pkg}
\alias{multibridge_pkg}
\title{\pkg{multibridge} package}
\format{
An object of class \code{package} of length 21.
}
\usage{
data(multibridge_pkg)
}
\description{
Metadata about a historic development version of the \pkg{multibridge} package.
}
\references{
TBD
}
\keyword{datasets}
